locked users only being able to delete posts that they own

+++ graphql-prisma/src/resolvers/Mutations.js
@@ -43,18 +43,19 @@ const Mutation = {
       token: jwt.sign({ userId: user.id }, "thisisasecret"),
     };
   },
-  async deleteUser(parent, args, { prisma }, info) {
-    const user = await prisma.exists.User({ id: args.id });
-    if (!user) {
-      throw new Error("Could not find user");
-    }
-    return prisma.mutation.deleteUser({ where: { id: args.id } }, info);
+  async deleteUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+
+    return prisma.mutation.deleteUser({ where: { id: userId } }, info);
   },
-  updateUser(parent, args, { prisma }, info) {
-    return prisma.mutationn.updateUser(
+  updateUser(parent, args, { prisma, request }, info) {
+    console.log(args.data);
+    const userId = getUserId(request);
+    console.log(userId);
+    return prisma.mutation.updateUser(
       {
         where: {
-          id: args.id,
+          id: userId,
         },
         data: args.data,
       },
@@ -79,7 +80,19 @@ const Mutation = {
       info
     );
   },
-  deletePost(parent, args, { prisma }, info) {
+  async deletePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    })
+
+    if(!postExists){
+      throw new Error("Unable to delete post")
+    }
+
     return prisma.mutation.deletePost({ where: { id: args.id } }, info);
   },
   updatePost(parent, args, { prisma }, info) {
